Moved cookies parsing to subpackage, added Get() method
