+++ utils.go
@@ -532,10 +532,3 @@ func NewFileUploadRequest(req FileUploadRequest) (*http.Request, error) {
 	return request, nil
 }
 
-// ParseCookieString parses string containing cookies and returns a slice of pointers to http.Cookie objects
-func ParseCookieString(cookie string) []*http.Cookie {
-	header := http.Header{}
-	header.Add("Cookie", cookie)
-	request := http.Request{Header: header}
-	return request.Cookies()
-}
